ZU-469 - Make event dispatcher support arrays
The event dispatcher casts all the values sent to it to string, which makes it impossible
to send arrays.

This adds a special case for arrays where we map each value in the array to a string if
it's not a scalar, and otherwise if it's scalar we leave it alone.

Similarly, other values are only cast to string if they are not scalar. This allows passing
non-strings to commands through the event dispatcher..